requirements.txt
aioredis==1.3.0
asgiref==3.2.2
asn1crypto==1.1.0
async-timeout==3.0.1
attrs==19.2.0
autobahn==19.10.1
Automat==0.7.0
backports.csv==1.0.7
CacheControl==0.12.5
cachetools==3.1.1
certifi==2019.9.11
cffi==1.12.3
channels==2.3.0
channels-redis==2.4.0
chardet==3.0.4
constantly==15.1.0
cryptography==2.7
cycler==0.10.0
daphne==2.3.0
defusedxml==0.6.0
diff-match-patch==20181111
dill==0.3.1.1
dj-database-url==0.5.0
Django==3.0.2
django-common-helpers==0.9.2
django-compat==1.0.15
django-crontab==0.7.1
django-extensions==2.2.3
django-heroku==0.3.1
django-redis==4.10.0
djangorestframework==3.10.3
docutils==0.16
et-xmlfile==1.0.1
firebase-admin==3.1.0
geographiclib==1.50
Geohash==1.0
geopy==1.20.0
google-api-core==1.14.3
google-api-python-client==1.7.11
google-auth==1.6.3
google-auth-httplib2==0.0.3
google-cloud-core==1.0.3
google-cloud-firestore==1.5.0
google-cloud-storage==1.20.0
google-resumable-media==0.4.1
googleapis-common-protos==1.6.0
grpcio==1.24.3
gunicorn==19.9.0
hiredis==1.0.0
httplib2==0.14.0
hyperlink==19.0.0
idna==2.8
incremental==17.5.0
jdcal==1.4.1
jieba==0.39
joblib==0.14.0
kiwisolver==1.1.0
matplotlib==3.1.1
msgpack==0.6.2
msgpack-python==0.5.6
ndg-httpsclient==0.5.1
numpy==1.17.2
odfpy==1.4.0
openpyxl==3.0.0
pandas==0.25.1
pbr==3.1.1
Pillow==6.2.0
protobuf==3.10.0
psycopg2==2.8.3
pusher==2.1.4
pusher-push-notifications==2.0.0
pyasn1==0.4.7
pyasn1-modules==0.2.7
pycparser==2.19
pycryptodome==3.9.6
PyHamcrest==1.9.0
PyJWT==1.7.1
PyNaCl==1.3.0
pyOpenSSL==19.0.0
pyparsing==2.4.2
python-dateutil==2.8.0
python-dotenv==0.10.3
python-geohash==0.8.5
pytz==2019.2
PyYAML==5.1.2
razorpay==1.2.0
redis==2.10.6
requests==2.22.0
rsa==4.0
scikit-learn==0.21.3
scipy==1.3.1
Send2Trash==1.5.0
service-identity==18.1.0
six==1.12.0
sklearn==0.0
South==1.0.2
sqlparse==0.3.0
tablib==0.13.0
Twisted==19.7.0
txaio==18.8.1
uritemplate==3.0.0
urllib3==1.25.6
whitenoise==4.1.4
xlrd==1.2.0
xlwt==1.3.0
zope.interface==4.6.0
fcm-django==0.3.4